Output of std::string.find() needs to be received in unsigned long (unsigned int works incorrectly, without a warning!).

+++ code/synthesis/TransformMachines2/Utils.cc
@@ -760,7 +760,7 @@ namespace casa{
       {
 	char *valStr=NULL;
 	std::string tt(name);
-	unsigned int pos;
+	unsigned long pos;
 	while((pos=tt.find(".")) != tt.npos)
 	  tt.replace(pos, 1, "_");
